requirements.txt
backcall==0.2.0
click==8.0.3
decorator==5.1.0
environs==9.3.5
flake8==4.0.1
Flask==2.0.2
ipdb==0.13.9
ipython==7.29.0
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
MarkupSafe==2.0.1
marshmallow==3.14.1
matplotlib-inline==0.1.3
mccabe==0.6.1
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.22
psycopg2-binary==2.9.2
ptyprocess==0.7.0
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.10.0
python-dotenv==0.19.2
toml==0.10.2
traitlets==5.1.1
wcwidth==0.2.5
Werkzeug==2.0.2
